Add workflow for static Linux build
These changes modify the OpenOCD Snapshot workflow to produce statically linked Linux binaries in addition to the Windows binaries. The release stage of the workflow has been split out into a separate job that aggregates the artifacts produced in Linux and Windows builds and adds them to a new release if a git tag exists.